% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsSNPCalculate.R
\name{findGeneInfo}
\alias{findGeneInfo}
\title{Helper}
\usage{
findGeneInfo(chrName, startPosition, endPosition)
}
\arguments{
\item{chrName}{the chromosome name of user input in the main function}

\item{startPosition}{the starting coordinate of user input in the main function}

\item{endPosition}{the end coordinate of user input in the main function}
}
\description{
load the retrieved information from the BioMart database according input
Note: the input should be valid since the preliminary check was done in the
main functions
}
