% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPFreqPlot.R
\name{SNPFreqPlot}
\alias{SNPFreqPlot}
\title{Generate an overview plot of SNP locations versus number of nucleotide
variants by gene range}
\usage{
SNPFreqPlot(chrName, start_position, end_position)
}
\arguments{
\item{chrName}{A integer value of class "numeric" indicating
the human chromosome 1-22, or a char either in 'X'or 'Y' indicating
human sex chromosome.}

\item{start_position}{A positive integer indicating the starting coordinate
of gene range.}

\item{end_position}{A positive integer indicating the end coordinate
of gene range.}
}
\value{
Returns a plot indicating the overview SNPs
}
\description{
A function that integrate the information from bioMart and plot the
SNP positions over the protein encoding range in
given chromosome, gene range (start coordinate, end coordinate).
}
\examples{
# Calculates percentage of nsSNPs in chromosome 1gene range 2321253 to 2391707
SNPplot <- SNPFreqPlot(chrName = 1,
                       start_position = 2321253,
                       end_position = 2391707)
SNPplot

}
\references{
Durinck, S., Spellman, P., Birney, E.,& Huber, W. (2009). Mapping identifiers
for the integration ofgenomic datasets with the R/Bioconductor package
biomaRt. \emph{Nature Protocols}, 4, 1184–1191.2.

Durinck, S., Moreau, Y., Kasprzyk, A., Davis, S., De Moor, B., Brazma,
A.,& Huber, W. (2005).BioMart and Bioconductor: a powerful link between
biological databases and microarray data analysis.\emph{Bioinformatics},
21, 3439–3440.

Pagès, H., Aboyoun, P., Gentleman, R., DebRoy, S. (2021). Biostrings:
Efficient manipulation of biological strings. R package version 2.62.0,
https://bioconductor.org/packages/Biostrings.

Pagès, H. (2017). SNPlocs.Hsapiens.dbSNP144.GRCh38: SNP locations for Homo
sapiens (dbSNPBuild 144). R package version 0.99.20.
}
