% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPFreqPlot.R
\name{SNPFreqPlot}
\alias{SNPFreqPlot}
\title{Generate a plot of SNP locations versus number of nucleotide
variants by gene range}
\usage{
SNPFreqPlot(chrName, startPosition, endPosition)
}
\arguments{
\item{chrName}{A integer value of class "numeric" indicating
the human chromosome 1-22, or a char either in 'X' or 'Y' indicating
human sex chromosome.}

\item{startPosition}{A positive integer indicating the starting coordinate
of gene range.}

\item{endPosition}{A positive integer indicating the end coordinate
of gene range.}
}
\value{
Returns a plot indicating the overview situation of SNPs
}
\description{
A function that integrate the information from bioMart and plot the
SNP positions over the protein encoding range in
given chromosome, gene range (start coordinate, end coordinate).
For better resolution, the genome range is limited within 200-nt length.
}
\examples{
# The used dataset of SNPs is default from SNPlocs.Hsapiens.dbSNP144.GRCh38 package
# The used dataset of Human genome is default from BSgenome.Hsapiens.UCSC.hg38 package

# Generate the SNP locations in chromosome 1 in region of 2321253 to 2321453
SNPFreqPlot(chrName = 1,
           startPosition = 2321253,
           endPosition = 2321453)

}
\references{
Durinck, S., Spellman, P., Birney, E.,& Huber, W. (2009). Mapping identifiers
for the integration ofgenomic datasets with the R/Bioconductor package
biomaRt. \emph{Nature Protocols}, 4, 1184–1191.2.

Durinck, S., Moreau, Y., Kasprzyk, A., Davis, S., De Moor, B., Brazma,
A.,& Huber, W. (2005).BioMart and Bioconductor: a powerful link between
biological databases and microarray data analysis.\emph{Bioinformatics},
21, 3439–3440.

Pagès, H., Aboyoun, P., Gentleman, R., DebRoy, S. (2021). Biostrings:
Efficient manipulation of biological strings. R package version 2.62.0,
https://bioconductor.org/packages/Biostrings.

Pagès, H. (2017). SNPlocs.Hsapiens.dbSNP144.GRCh38: SNP locations for Homo
sapiens (dbSNPBuild 144). R package version 0.99.20.

Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag
New York. ISBN 978-3-319-24277-4, https://ggplot2.tidyverse.org.
}
